+++ src/Brightree.php
@@ -4,6 +4,7 @@ namespace Nickcheek\Brightree;
 
 use SoapClient;
 use Nickcheek\Brightree\Service\Patient;
+use Nickcheek\Brightree\Service\Doctor;
 use Nickcheek\Brightree\Service\Document;
 use Nickcheek\Brightree\Service\DocumentManagement;
 use Nickcheek\Brightree\Service\CustomField;
@@ -35,6 +36,11 @@ class Brightree
     public function Insurance()
     {
     	return new Insurance();
+    }
+
+    public function Doctor()
+    {
+    	return new Doctor();
     }